Fixes closing of Cache and Policy to be concurrent safe
Before this change, calling Close on the cache while a Set call is in progress could occasionally result in a panic when the Set method tried to write to the (now closed) `setBuf` channel. This commit changes Cache and Policy to keep references to the written-to channels in an `internals` struct, which is nil-ed out on close and checked in each call site, ensuring that nothing is attempting to write to a closed channel.